Changed wake-up-db.ts back to .js so it can be executed by the node process during CI
